Set architectures property to avr
This library is AVR specific so the previous wildcard value of the architectures property in library.properties is incorrect.

This change will cause this library's examples to appear under the File > Examples > INCOMPATIBLE > EEPROMex menu when a non-avr board is selected, which can provide a valuable clue to the user why the library isn't working with that board. The Arduino IDE also provides a library mismatch warning but unfortunately this does not currently appear when compilation fails due to an included file not being found, which is the outcome of attempting to compile for a non-AVR board. This issue will hopefully be resolved in a future version of the Arduino IDE.